:recycle: 유저 코드 관련 기능을 모두 분리
