+++ src/features/user-code/ui/friendSearch.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect, useCallback } from "react";
 import styled, { keyframes } from "styled-components";
-import { searchFriends } from "../../features/social/api/friends";
-import { useAuth } from "../../shared/lib/context/AuthContext";
+import { useAuth } from "../../../shared/lib/context/AuthContext";
 import { useQuery } from "@tanstack/react-query";
+import { findUserByCode } from "../api/userCodeApi";
 
 // Friend 인터페이스 정의
 interface Friend {
@@ -97,7 +97,8 @@ const FriendSearch: React.FC<FriendSearchProps> = ({ onClose }) => {
         queryKey: ['friends', 'search', query],
         queryFn: async () => {
             if (!user || !query.trim()) return [];
-            return searchFriends(user.id, query);
+            const response = await findUserByCode(query);
+            return [response.data];
         },
         enabled: !!user && query.trim() !== "",
         staleTime: 1000 * 60, // 1분 동안 캐시 유지
@@ -118,7 +119,7 @@ const FriendSearch: React.FC<FriendSearchProps> = ({ onClose }) => {
             )}
             {results.length > 0 && (
                 <ResultList>
-                    {results.map((friend) => (
+                    {results.map((friend: Friend) => (
                         <ResultItem key={friend.id}>
                             {friend.username}
                         </ResultItem>
